[UPD] refactor code
