+++ src/utils/helper.js
@@ -1,0 +1,3 @@
+export const rand = (max, min) => {
+  return Math.floor(Math.random() * (max - min) + min);
+}
